fixed task2
