--- task1/src/main/java/com/shalkevich/interfaces/Table.java
@@ -1,10 +1,0 @@
-package com.shalkevich.interfaces;
-
-/**
- * Interface for table presentation.
- */
-
-public interface Table {
-    void print();
-
-}
