+++ task1/src/main/java/com/shalkevich/service/Table.java
@@ -1,0 +1,9 @@
+package com.shalkevich.service;
+
+/**
+ * Print multiplication table.
+ */
+public interface Table {
+    void print();
+
+}
